# User/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация к серверу User",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/send-mail-for-code-confirmation": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Отправка письма с кодом подтверждения",
                "description": "Метод отправляет письмо с кодом подтверждения на указанный email.",
                "operationId": "18ba9c802fbad2cdac33fde59f762141",
                "requestBody": {
                    "description": "Данные для отправки письма",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email пользователя, на который будет отправлен код подтверждения.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "kivanov032@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Письмо успешно отправлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус операции.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Сообщение об успешной отправке.",
                                            "type": "string",
                                            "example": "Сообщение успешно доставлено пользователю."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке валидации.",
                                            "type": "string",
                                            "example": "Поле email обязательно для заполнения."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле email обязательно для заполнения."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус операции.",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Сообщение об ошибке.",
                                            "type": "string",
                                            "example": "Ошибка отправки письма."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/confirm-code": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Подтверждение кода",
                "description": "Метод проверяет код подтверждения и возвращает статус успеха или ошибки.",
                "operationId": "75506d25e02c5cd0a64d35e34863a9c9",
                "requestBody": {
                    "description": "Код подтверждения",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "input_code"
                                ],
                                "properties": {
                                    "input_code": {
                                        "description": "Код подтверждения. Должен быть 6-значным числом.",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное подтверждение",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус успешного подтверждения.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Сообщение об успешном подтверждении.",
                                            "type": "string",
                                            "example": "Код подтверждения верен."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка подтверждения",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус ошибки.",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Сообщение об ошибке.",
                                            "type": "string",
                                            "example": "Неверный код подтверждения."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке валидации.",
                                            "type": "string",
                                            "example": "Поле input code должно быть не меньше 6 символов."
                                        },
                                        "errors": {
                                            "properties": {
                                                "input_code": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле input code должно быть не меньше 6 символов."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке.",
                                            "type": "string",
                                            "example": "Произошла внутренняя ошибка сервера."
                                        },
                                        "error": {
                                            "description": "Детали ошибки.",
                                            "type": "string",
                                            "example": "Сообщение об ошибке"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {your_token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}